Move mongodb connection adapter from common to network.storage.statestore
